LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/clients-matchall.https-expected.txt



FAIL Test Clients.matchAll() assert_array_equals: expected property 1 to be false but got true (expected array ["visible", false, "https://localhost:9443/service-workers/service-worker/resources/blank.html?clients-matchAll#2", "window", "nested"] got ["visible", true, "https://localhost:9443/service-workers/service-worker/resources/blank.html?clients-matchAll#2", "window", "nested"])
